pyproject.toml
[project]
name = "comuniveci-post-service"
version = "0.1.0"
description = ""
authors = [
    {name = "JulianAVG64",email = "julvargasgu@unal.edu.co"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)"
]

[tool.poetry]
packages = [{include = "comuniveci_post_service", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
